Local definitions of geometry types in Nullable don't work
